ai-services/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.0
openai==1.14.3
httpx==0.27.0
python-dotenv==1.0.1
langchain==0.1.12
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1
pymongo==4.6.2
redis==5.0.3
python-multipart==0.0.9
Pillow==10.2.0
detoxify==0.5.2
